requirements.txt
# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database Drivers
psycopg2-binary>=2.9.0
pymongo>=4.3.0
neo4j>=5.8.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Benchmarking
# hyperfine>=1.16.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Development & Testing
pytest>=7.2.0
black>=23.1.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
